+++ CHANGELOG.rst
@@ -17,6 +17,7 @@ Changed:
 
 * refactored main workflow with more readible organization #614
 * refactored conda envs within container to be on base and container definition is uncoupled #759
+* QC metric validation is performed across all the samples #800
 
 Removed:
 ^^^^^^^^